inflate: fix MSVC compiler warnings

+++ inflate.c
@@ -473,7 +473,7 @@ local int updatewindow(z_streamp strm, const Bytef *end, unsigned copy) {
 /* Restore state from registers in inflate() */
 #define RESTORE() \
     do { \
-        state->wnext = put - (state->window + state->wsize);\
+        state->wnext = (unsigned)(put - (state->window + state->wsize));\
         strm->avail_out = left; \
         strm->next_in = next; \
         strm->avail_in = have; \
@@ -916,7 +916,7 @@ int ZEXPORT inflate(z_streamp strm, int flush) {
             copy = state->length;
             if (copy) {
                 unsigned char *end = state->window + (state->wsize * 4);
-                unsigned long diff = end - put;
+                unsigned diff = (unsigned)(end - put);
 
                 if (copy > have) copy = have;
                 if (copy > diff) {
@@ -924,7 +924,7 @@ int ZEXPORT inflate(z_streamp strm, int flush) {
                         RESTORE();
                         window_output_flush(strm);
                         LOAD();
-                        diff = end - put;
+                        diff = (unsigned)(end - put);
                     }
                     if (copy > diff) copy = diff;
                 }
@@ -1175,7 +1175,7 @@ int ZEXPORT inflate(z_streamp strm, int flush) {
                 /* fallthrough */
             case MATCH: {
                 unsigned char *end = state->window + (state->wsize * 4);
-                unsigned long buf_left = end - put;
+                unsigned buf_left = (unsigned)(end - put);
                 copy = state->length;
 
 		RESTORE();
@@ -1184,7 +1184,7 @@ int ZEXPORT inflate(z_streamp strm, int flush) {
                         /* relies on RESTORE() above with no changes to those vars */
                         window_output_flush(strm);
                         LOAD();
-                        buf_left = end - put;
+                        buf_left = (unsigned)(end - put);
                     }
                     if (copy > buf_left) copy = buf_left;
                 }
